1.3.0 Move pack_aws types to types.go
