+++ s3packs/pack_aws/iterator.go
@@ -4,27 +4,12 @@ import (
 	"os"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/feature/s3/manager"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"github.com/orme292/s3packer/conf"
 	"github.com/orme292/s3packer/s3packs/objectify"
 	"github.com/orme292/s3packer/s3packs/provider"
 )
-
-type AwsIterator struct {
-	provider *conf.Provider
-	svc      *manager.Uploader
-	fol      objectify.FileObjList
-	stage    struct {
-		index int
-		fo    *objectify.FileObj
-		f     *os.File
-	}
-	group int
-	err   error
-	ac    *conf.AppConfig
-}
 
 func AwsIteratorFunc(ac *conf.AppConfig, fol objectify.FileObjList, grp int) (iter provider.Iterator, err error) {
 	return NewIterator(ac, fol, grp)
